decouple parsing the yaml manifest from the declarative source implementation
